Refactor CreateTranslator IPC
Currently the CreateTranslator IPC between the browser process and the
on device translation service accepts a pending receiver of Translator,
and the result callback will be called with a bool `is_sucessful`.
  CreateTranslator(string source_lang,
                   string target_lang,
                   pending_receiver<Translator> receiver)
      => (bool is_sucessful);
This is a bit redundant. This can be simplified as follows:
  CreateTranslator(string source_lang,
                   string target_lang)
      => (pending_remote<Translator>? translator);

This CL also changes the CreateTranslator IPC between the renderer
process and the browser process as follows:
  CreateTranslator(
    pending_remote<TranslationManagerCreateTranslatorClient> client,
    TranslatorCreateOptions options);
The IPC accepts a TranslationManagerCreateTranslatorClient's pending
remote. This is intended to implement the request abortion in future.
This is following the similar logic of AIManager's Create* IPCs.

Bug: 374631433
Change-Id: Iad37e3361fe66373bc3a45d07a70f6410a3ae3a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5964098
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Ming-Ying Chung <mych@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374519}
NOKEYCHECK=True
GitOrigin-RevId: 92cd676f3af9be0bda37740c934cc7b2d0584c95