Migrate transferBackFromGPUTexture_context_lost to canvasPromiseTest
The original test was lacking worker test coverage because the
forceLoseCanvasContext internal API wasn't implemented for workers.
canvasPromiseTest runs the test on both the main thread and in a
worker. For this to work, this CL adds a worker implementation for
internals.forceLoseCanvasContext.

Bug: 374263392
Change-Id: Id241d5acc5532e034c8bb8c066f33a3a4a3a36b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5939737
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374610}
NOKEYCHECK=True
GitOrigin-RevId: b115dc11e498e17d7eac04be39dbd22a8a155eb2