Clean up integrity checking code.
We check a resource's integrity disposition in a few places, with
somewhat inconsistent formatting (checking "!passed" vs "failed",
for instance). To make it simpler to add some functionality here,
this patch makes a few small changes to align these callsites,
and centralize the decision about whether or not to force
integrity checks for cached/prefetched resources.

No change in behavior is intended; the checks should have the same
outcomes as in the status quo.

Bug: 375343417
Change-Id: Iea10d3f730cc60359790070fcb06ce62acb97f48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5965057
Reviewed-by: Yoav Weiss (@Shopify) <yoavweiss@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374688}
NOKEYCHECK=True
GitOrigin-RevId: 2f7e8821db5faa4bd9a3592fa96fb3d5cef45a7c