[A11y] In named slot assignment mode, remove a11y subtrees if they change TreeScope
Mirror the logic of the manual assignment case, to the auto/named
case, by removing the subtree when the tree scope of a slotted
element changes. Thus if an element that previously had an AXObject is
slotted to an <slot> that does not have an AXObject (e.g. is hidden),
then it will not break out checks that every AXObject has a parent.

Fixed: 372055504,375768987
Change-Id: Ib4984b02709660e3ce1e18ceb637dd48e8fe8eb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5968253
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374771}
NOKEYCHECK=True
GitOrigin-RevId: dd03e8c56b93e4c4904ce8205e550bda69a163fb