Add test for changing drawing styles in TextClusters
Following the discussion on the WhatWG issue for Enhanced TextMetrics,
it was agreed that TextClusters should be rendered with values defined
in CanvasTexDrawingStyles as they were when ctx.measureText() was
called, with the exception of text align and baseline if they are passed
as options to TextMetrics::getTextClusters() (see:
https://html.spec.whatwg.org/#canvastextdrawingstyles).

This means that even if any of these values changes between the moment
that ctx.measureText() was called, and the time they are rendered with
ctx.fillTextCluster(), the styles will be preserved as they were when
the text was measured.

A previous test already checks if that is the case for the font. This CL
adds a test that checks this for letterSpacing, wordSpacing, fontKerning,
and fontVariantCaps. Some drawing styles require more specific set ups
(fontStretch and textRendering), and will have tests added in future CLs.

Bug: 341213359
Change-Id: I4b123633b3e4951385145596388cf5b192221465
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5948244
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374809}
NOKEYCHECK=True
GitOrigin-RevId: c3f183826bedc841f769030fd28deadd7a577eed