Unify ::search-text:current and "active match" terminology
A decision has been made to use :current to identify the
currently-focused text match:
https://github.com/w3c/csswg-drafts/issues/10212

However so far find-in-page has used the term "active match" for the
same concept. Since that is used a lot throughout the codebase, it is
more practical to use "active match" too for the new :text-search pseudo
element, except where we are dealing with CSS styling.

Bug: 369223564

Change-Id: I76c1964b2e917ffcd9ff1edc678eb00956fdf5a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5886449
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Rob Buis <rbuis@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1374883}
NOKEYCHECK=True
GitOrigin-RevId: 62c404c5000c28de99717d8072bb952065a28f10