Remove `NOTREACHED` in `PaintLineBoxes`
This patch removes a `NOTREACHED` that hits when paint finds
needs-layout objects.

This `NOTREACHED` is causing a P1 stability issue as it is
being turned to crashes. It shouldn't still happen, but
`DCHECK`-enabled build should have hit checks earlier, that
are closer to the actual causes, such as `AssertLaidOut()`.

This redundant `NOTREACHED` helped to clarify the cause is
likely in layout, but not as much important as to cause a P1
stability issue.

Fixed: 375000982
Change-Id: I1bd23c60914923458a3af08b7fc44e15f4ae1e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5973362
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1375064}
NOKEYCHECK=True
GitOrigin-RevId: 4b2f885bcd2cbec79adc8cd0810026c77cabc614