[bindings][fastapi] Create HandleScope only when necessary
When the fallback was removed from fast-API callbacks, an unconditional
HandleScope was introduced to handle potential handles created when
throwing an exception. With this CL, a HandleScope is created only when
the API function may throw an exception, or when one of the parameters
is a HeapObject.

Bug: 41492790
Change-Id: Ie0b9668ff7f78a2d380e96bf42d7ef1c5131abd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5971873
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1375377}
NOKEYCHECK=True
GitOrigin-RevId: 84e8d729edbab262c0ca48625646980dc431f556