Fix incremental progress bar in run_web_tests.
By default (no --verbose or --debug-rwt-logging), run_web_tests is
supposed to show a ninja-style incremental progress bar while
running (i.e., it prints "[1/10000] foo/bar.html" and overwrites
that line as we progress through the tests).

The line is printed to stderr, and in Python 2, stderr is unbuffered,
meaning that lines are immediately flushed and printed. In Python 3,
stderr changed to being buffered, meaning that non of the updates
get printed until something else needs to be printed (like a test
failure, or the completion of the tests). Which defeats the point.

This CL explicitly calls out to flush stderr after printing the
progress update.

Change-Id: Ia7f1c97b9d514982cae1fa06613d2dbd72c254ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5972803
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1375438}
NOKEYCHECK=True
GitOrigin-RevId: 3e00a1bb42d65b034d97e45c99967d6df17bb878