[A11y] Add null check in case aria-hidden ancestor is pruned
Correcting aria-hidden and/or focus could cause an object to be
removed, e.g. if it turns out to be outside of an aria-modal on Mac.
This is a follow-up to yesterday's fix on the same bug.

Bug: 375669433
Change-Id: I265f4b0f5e2507d88ed89b049ee685dde9fe2da5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5978848
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Tzarial <zork@chromium.org>
Commit-Queue: Tzarial <zork@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1375814}
NOKEYCHECK=True
GitOrigin-RevId: adb66b55d182d374ae084db67fb995bc33c6d855