Fenced frame: refactor navigate ancestor tests for headless mode.
User activation does not work in headless mode when dealing with new
windows. This CL refactors our navigate ancestor tests to run the test
in the main window. This means that the failure condition will now be
different. Rather than an assert being hit, if the test fails, it will
now time out, as the top-level page will have been navigated away.

This also simplifies the setup to use remote context executors instead
of a series of various helper files. If this causes flakes down the
line, this will be reverted.

Change-Id: I492bd524ad755c8e310264804650ff213a027d80
Bug: 366257368
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5974698
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Andrew Verge <averge@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1375978}
NOKEYCHECK=True
GitOrigin-RevId: bcb037bb3b68ef8a7eb0365b89ca7e3b5594a8e2