Make GetCacheIdentifier() respect GetSkipServiceWorker().
Since the current GetCacheIdentifier() ignores GetSkipServiceWorker(),
GetCacheIdentifier() returns ServiceWorkerId even if GetSkipServiceWorker()
is true if the ServiceWorker has a fetch handler.

To make the isolated world respected as an isolated world, the cache
identifier should not be shared with a page under a ServiceWorker control.

Bug: 372512079, 373263969
Change-Id: Idd2d8900f2f720e0a4dc9837e2eb56474c60b587
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5961018
Reviewed-by: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376006}
NOKEYCHECK=True
GitOrigin-RevId: 75f322ad1f64c0bc56fa77ab877b48d72cdb903c