[A11y] Handle when document is marked dirty late in the process
This condition actually occurs only on Mac, because it is the only
platform that needs to use MarkDocumentDirty() when aria-modal dialogs
are created or destroyed. There is no real additional cost to keep the
code for all platforms.

Will do test as a follow-up. Repro is to load pepboys.com and
continually click on "Your store" on the right side of the toolbar.

Fixed: 372508699,375669433
Change-Id: I524479e035d27e6118a53fdccef77fc9ad5d7dcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979112
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376030}
NOKEYCHECK=True
GitOrigin-RevId: f2eb83b50beff612ec481f600cb0d9cfe05dedfd