[Commerce] Update commerce internals page to use Lit
This patch updates the commerce internals page to use Lit instead of
manually manipulating the DOM through vanilla JS. This is possible now
that Lit is supported on iOS and Android. The functionality of the page
is the same with minimal UI improvements. Migrating to Lit will make it
easier to add functionality in the future and allow for the use of other
Lit elements (like cr_components) if desired.

Low-Coverage-Reason: OTHER This is an internal commerce page intended
for debugging.

Bug: 376112788
Change-Id: Idfe9a7873f07cf2e15988decb9ba480d5e76c7fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5974291
Reviewed-by: Ayman Almadhoun <ayman@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Chase Hartsell <chazzy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376038}
NOKEYCHECK=True
GitOrigin-RevId: 10a2858ce08d28f0b80b222d56fcf740c899c7ee