[blink] Remove redundant calls to ElapsedTimer in parsing.
base::ElapsedTimer::Elapsed() calls clock_gettime(), which is mostly
fast, but can be slower on some ARM devices, due to CPU errata that
forces the kernel to take a slow path in the VDSO.

From a local speedometer profile, a large contributor to the calls on
the renderer main thread is ElapsedTimer. This CL removes the redundant
ones.

Bug: 371033180
Change-Id: I9173186dda890e1f602b5b521eb2afacb97b4780
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979852
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376247}
NOKEYCHECK=True
GitOrigin-RevId: f0001ea156e4c19de472ff0a50c8342c1144110f