Remove old error message in GetGainmapInfoAndData
The error message indicates a bug in libavif (decoder_->gainMapPresent
is true but decoder_->image->gainMap is a null pointer). The new code
checks a different condition (the image does not have a gain map), so it
should not log this error message.

Change-Id: Id752d0913e95ca9af06a1dd40a42ef7387aae4c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5980878
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Wan-Teh Chang <wtc@google.com>
Cr-Commit-Position: refs/heads/main@{#1376498}
NOKEYCHECK=True
GitOrigin-RevId: 30f54b467d3530f782492e59fabbf7bffce31c74