renderTime: always coarsen up, and delay adding the entry
This is a more robust way to ensure that render times don't
end up being "in the future".

The render time is aligned to the upper nearest multiple of 4ms.
The current timestamp is also aligned in the same way, and the
result of that is used to post a delayed task that ends up
adding the entry itself.

Also adjust some metric tests epsilons.

Bug: 376029956
Bug: 376096087
Bug: 376044090
Bug: 376075878
Bug: 376177179
Bug: 376175673

Update metric test epsilon

Change-Id: If6adc9562d7ca89490371cee5d847c070e1efabb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5975665
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376574}
NOKEYCHECK=True
GitOrigin-RevId: 59c59f02d180e10de20e99fba8da5ed4828ba076