[Grid] Adding IsGridRepeatValue to simplify logic
Addresses a TODO to combine these checks at the CSSValue
level. This uses a simple OR instead of an inheritance
hierarchy as the TODO suggests, as this follows the patterns
used in CSSValue. This approach is also less code and
potentially more performant than a virtual method and cast.

No behavioral changes are expected.

Change-Id: I49c373a84323b065a0795738bdd783fed118c160
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5981179
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1376595}
NOKEYCHECK=True
GitOrigin-RevId: cedbbdeeba92126558e8d894ecbf938790be90ed