[WebRTC] Leak audio processor in WebRtcAudioSink::Adapter if Post fails
Some tests are crashing because PostCrossThreadTask fails and an
object is destroyed on the wrong thread, resulting in a crash.

This CL tries to fix this problem by leaking the object instead of
destroying it on the wrong thread when PostCrossThreadTask fails.

Bug: 350587808
Change-Id: I831e21d2d0f7a1bf1dfbedb4e180986c67535abc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5872533
Reviewed-by: Evan Shrubsole <eshr@google.com>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376834}
NOKEYCHECK=True
GitOrigin-RevId: 0083a07dd690e3d4293845c6243e8751ec5b9a8f