[nesting] Improve WPT coverage for top-level '&' specificity
The '&' selector, when it has no parent selector to draw from,
has a specificity of zero [1].

Blink already does the right thing. This CL just adds a new WPT
to enforce that behavior, and adjusts an existing tests which
expected specificity to be added.

[1] https://github.com/w3c/csswg-drafts/issues/10196

Bug: 343794754
Change-Id: I3bfdba8a5db66be2d2cdd6f1b2f41b86aadbab9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5982912
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376861}
NOKEYCHECK=True
GitOrigin-RevId: 7a861942f5c428164a0725042bc22cf84eb83575