Add ::select-arrow pseudo element
This pseudo element is added to allow <select> elements to have a
standardized and customizable drop down icon.

Spec definition:
```
Select dropdown icon: the '::select-arrow' pseudo-element

The '::select-arrow' pseudo-element is only generated when the
originating element is a select element with base appearance.

'::select-arrow' is a tree-abiding pseudo-element. It accepts all
properties. It inherits from its originating element.

'::select-arrow' generates a box as if it was an child of its
originating element, after any boxes generated by the '::after'
pseudo-element, with content as specified by 'content'.
```

Defined in:
https://github.com/w3c/csswg-drafts/pull/10986

Bug: 369319576
Change-Id: Iaea82a130aecdaeacb07112416a365aa8d3349af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5973559
Commit-Queue: Traian Captan <tcaptan@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376907}
NOKEYCHECK=True
GitOrigin-RevId: 58d54c80b37fccdab768379e93ddc19e7b6965be