[A11y] Repair names for tabbable elements only if focused
This fixes a performance regression that occurs with the
KeyboardFocusableScrollers feature.

In some cases we repair accessible names for elements that
are keyboard-focusable but are missing a name. That way, the user
will hear something other than silence as they tab to the object.
However, it's not actually necessary to repair the name for every
tabbable object, just the one that receives focus. In fact,
it's arguably harmful to included it for every object as a repaired
name from contents can create very large names in the case of regions,
and only serves to cause redundant announcements when visited with
the virtual buffer.

In addition, our name repairs should not depend on Element::IsKeyboardFocusable() which can be very slow, as it needs
to check whether there are keyboard-focusable descendants before
returning true for a scrollable region (if there are
keyboard-focusable descendants, it will return false). This is not
an issue where it's used for tab navigation, but if we are making
the call many times in a row in can be very slow. That was the case
when walked ancestors for every DOM change to determine whether any
ancestor has a name change. Each call to SupportsNameFromContents()
could potentially need to call the slow IsKeyboardFocusable()
method for any scrollable region.

To fix both issues, this CL removes a11y use of
Element::IsKeyboardFocusable() and instead computes the name
for scrollable regions only when they are focused.

In addition, this CL removes AXObject::IsKeyboardFocusable() which
already was not being used.

Bug: 375185771, 375937552, 375352863
Change-Id: Ib90f2175b74c2149109b14205bdac3ee7388972a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979028
Reviewed-by: Jacob Francis <francisjp@google.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376917}
NOKEYCHECK=True
GitOrigin-RevId: 41644220e3f20971fd8b5e24430a0753a8cbac02