[PEPC] Track occluder node ID if the element is occluded
This CL to help us figure out why an element isn't activated. When
something's overlapping the element, it can be tough to tell what's
going on, especially if the occluder is transparent and generated from
some frontend frameworks.

Not in this CL, we will consider also keeping track of any visual
effects that might be affecting the activation status of the element,
if necessary.

Bug: 363981300
Change-Id: I5d6f420dd49a1ac31e6026b5ff7545a07396a836
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5836863
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Thomas Nguyen <tungnh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377098}
NOKEYCHECK=True
GitOrigin-RevId: c4e5511fc37ccff015684edcfa462e31310b7a78