IDB: Report NotReadableError for NotFound large value read errors
As a followup to crrev.com/c/5831673 and
https://github.com/w3c/IndexedDB/pull/430, this CL updates the name and
message of the DOMException returned when unwrapping of large IndexedDB
values fails due to the data files being missing from the disk.

The intention is to convey the cause and unrecoverability of the error
clearly to developers so that they can add appropriate error handling.

Change-Id: Ieefc4a8b2e9d6bfe9823bcb88394e438db322162
Bug: 362123231
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5971136
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1377584}
NOKEYCHECK=True
GitOrigin-RevId: 3fa6ab8b21426961aab282b9279372da2fb5f91f