Update cookies to have correct hasCrossSite ancestor values in test
Updating this test to use the correct values for hasCrossSiteAncestor. This is needed because this [CL](https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5738723) made changes to how the PartitionKey is created in Cookie.ts and recorded in NetworkRequests.ts. The expected values were not updated in this test and shortly after the CL landed, this test became flaky.

Bug: 357220234, 376854707
Change-Id: I369c383269443fca7a6db4e05dcf2002bda0dbc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5988713
Reviewed-by: Sandor «Alex» Major <sandormajor@google.com>
Commit-Queue: Aaron Selya <selya@google.com>
Cr-Commit-Position: refs/heads/main@{#1377883}
NOKEYCHECK=True
GitOrigin-RevId: 4b52122e15eb3021d56b71b8a728f4abc73ea367