[v8 code cache] Fix the wpt test added in https://chromium-review.googlesource.com/c/chromium/src/+/5964918
The problem with the test is that one of the fetch() calls
ends up creating the code cache. The fetch response has its
own response time, independent of the response time of the
resource in the HTTP cache.

Thus, we always get a response time mismatch when trying to
use the code cache. This was not the intention of the test.

The fix uses fetch() for targeted cache busting only at the
last step, so it won't interfere with code cache creation.

Bug: 373703283
Change-Id: I5766af7314fd6ffa920fe5d68c704c0611613f72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979864
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1378159}
NOKEYCHECK=True
GitOrigin-RevId: ef72ded5344a26b9e5ff1cedc9146d3c7c6a86a5