[Blink] Migrate HTMLCanvasElement from checking bridge when snapshotting
The fundamental reason that it checks the bridge is that the bridge
being presents means hibernation is possible, in which case the snapshot
should come from the hibernation handler rather than the resource
provider. This CL changes the code to directly check whether the
hibernation handler is present.

Bug: 40280152
Change-Id: I015e840d7ab4c421deea842ae44a18ac4b2b4122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5987894
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1378176}
NOKEYCHECK=True
GitOrigin-RevId: 847625a0928e4dd96bfbb3673f3e4a605410f187