Rename ConsumeDeclarationList to ConsumeBlockContents
The current editor's draft of the CSS Syntax spec no longer defines
the algorithm "parse a list of declarations", and instead relies on
"consume a block’s contents" for both declarations and child rules.

This seems like a better name for Blink as well, so we change
the function name to match.

Change-Id: Ifa931e35d644242bd774aa007a25bc4b202cb743
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5987671
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1378207}
NOKEYCHECK=True
GitOrigin-RevId: 46ef03078a74aa72df4db2d37bc66fdc31967bf6