Clean-up outdated permission-element test expectations
The expectations in disable-site-isolation-trials are no longer needed
as the underlying crash was fixed.

The expectations in TestExpectations are incorrect as we expect the
feature to be enabled since it's in the experimental state. If this
changes these tests should reflect that instead of silently allowing
everything.

Fixed: 375237147
Change-Id: Iedeaeacec804ab5a37a0ae1de236d0abe9bf69a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5982614
Reviewed-by: Florian Jacky <fjacky@chromium.org>
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1378312}
NOKEYCHECK=True
GitOrigin-RevId: df21cb1e2dea6ae37ad18298be37ec039d73b30e