Don't update LCP for video elements switching from poster.
A video element whose rendered contents are a poster image will have a
different MediaTiming than when it's contents are the video itself.
However, that different object will cause the MediaRecordId to be
different, and so the same element can create two distinct LCP candidate
records.

This change forces the MediaRecordId to not consider the MediaTiming
object when calculating the hash for a video element.

Bug: 330202431
Change-Id: I56b8c8d54c1a3d64138f8de529fe15695b1df952
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5970217
Reviewed-by: Hao Liu <haoliuk@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1378586}
NOKEYCHECK=True
GitOrigin-RevId: 2a184355d2a123af46a626979d43d7a991ab679e