[canvas] Use [NoAllocDirectCall] for CanvasRenderingContext2D::stroke
This is a partial reland of https://crrev.com/c/5813787. That CL
was reverted because the change to `fill()` was incorrect. However,
the performance improvements of that CL actually come from `stroke()`,
So with this CL we just use the V8 fast API for `stroke()`. Since
`stroke()` does not have any optional parameters, using the fast API
should not cause problems as with the optional parameter of `fill()`.

Change-Id: I24f9ac7d862ab419fa8215e6e25c323508851609
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5992800
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1378854}
NOKEYCHECK=True
GitOrigin-RevId: 8cd5a88245a4d5d440d15c921ea9220adaa709ca