PrivateAggregation: Add ReportPoller to WPTs for polling reports
This CL adds a class named `ReportPoller` to help Private Aggregation
WPTs correctly poll for reports. It only reworks the polling logic in
shared-storage-filtering-id-sends-report.https.window.js, but the same
principle applies to the other tests.

Prior to this CL, a few of these Shared Storage / filtering ID tests
called `pollReports()`, but accidentally forgot to set the timeout to
6000 ms. Consequently, they were falling back to the default of 5000 ms,
which leaves no wiggle room relative to the 5000 ms timeout chosen by
our Shared Storage implementation. This test defect might be responsible
for some flaky failures we've seen. Especially since some tests
subsequently poll for debug reports with a timeout of 50 ms!

The changes in this CL bring the runtime from ~25 seconds to ~50 seconds
(shared-storage-filtering-id-sends-report.https.window.js). I think this
is not due to an error in this CL, but rather because many tests have
not been polling for as long as they should have been.

Bug: 362901607
Change-Id: I627cd0ff705105cdbdba32895e2606af36d5a81b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5990990
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Alex Turner <alexmt@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1379133}
NOKEYCHECK=True
GitOrigin-RevId: 9997cf774a43af4380ff3a66fdaab1a2c16aebe4