Fix Canvas WebDXFeature use counters names
In the cl https://chromium-review.googlesource.com/c/chromium/src/+/5950129,
I missed the requirement that the counter name has to match with the
one in https://github.com/web-platform-dx/web-features.

Update the names to match and remove the ones are not required.

Change-Id: I2449e1edca61378d132c26a267ce534548d18966
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5988613
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Yi Xu <yiyix@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Andres Ricardo Perez <andresrperez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1379878}
NOKEYCHECK=True
GitOrigin-RevId: ef7b42a31f3dc894fa17ba110fae7800551bfb9e