Fix datetimelocal test to use eventSender tab
We previously used focus() because eventSender.keyDown('Tab') was
causing a crash with the LayoutTree update. That is no longer true so
we can move the test back to using eventSender.

Note that for this test, we need 3 tabs to get to the time column
of the datetime picker (focus navigate clear -> today -> time).

Change-Id: Idcf229b11345789af4f17a962b6479768c3f70d1
Fixed: 40679444
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6000993
Auto-Submit: Di Zhang <dizhangg@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1379941}
NOKEYCHECK=True
GitOrigin-RevId: 7bd055da5db18b2410a232126daa52891771f6cd