SelectAudioOutput() UI
This CL implements the browser-side UI and logic for the
selectAudioOutput() API. If necessary, a UI dialog presents available
audio output devices for user selection. The chosen device (or an
error) is then relayed back to the renderer via a callback,
completing the API call and providing the selected MediaDeviceInfo.

Bug: 372214870
Change-Id: I0f44ddfe0eed7b68e72655715d7e26932b2a8caa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5890611
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Palak Agarwal <agpalak@chromium.org>
Commit-Queue: Olha Sirikova <sirikovaolha@google.com>
Cr-Commit-Position: refs/heads/main@{#1380481}
NOKEYCHECK=True
GitOrigin-RevId: a6c4d523c60f1e3f332c234b3766773b0ba9fc02